refactor: move create attendance
