+++ src/main/java/org/cotato/csquiz/domain/attendance/service/AttendanceAdminService.java
@@ -1,7 +1,6 @@
 package org.cotato.csquiz.domain.attendance.service;
 
 import jakarta.persistence.EntityNotFoundException;
-import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -12,17 +11,11 @@ import lombok.extern.slf4j.Slf4j;
 import org.cotato.csquiz.api.attendance.dto.AttendanceRecordResponse;
 import org.cotato.csquiz.api.attendance.dto.AttendanceStatistic;
 import org.cotato.csquiz.api.attendance.dto.GenerationMemberAttendanceRecordResponse;
-import org.cotato.csquiz.common.error.ErrorCode;
-import org.cotato.csquiz.common.error.exception.AppException;
-import org.cotato.csquiz.domain.attendance.embedded.Location;
 import org.cotato.csquiz.domain.attendance.entity.Attendance;
-import org.cotato.csquiz.domain.attendance.entity.AttendanceRecord;
 import org.cotato.csquiz.domain.attendance.enums.AttendanceResult;
 import org.cotato.csquiz.domain.attendance.enums.AttendanceType;
-import org.cotato.csquiz.domain.attendance.repository.AttendanceRecordRepository;
 import org.cotato.csquiz.domain.attendance.repository.AttendanceRepository;
 import org.cotato.csquiz.domain.attendance.util.AttendanceExcelUtil;
-import org.cotato.csquiz.domain.attendance.util.AttendanceUtil;
 import org.cotato.csquiz.domain.auth.entity.Member;
 import org.cotato.csquiz.domain.auth.service.MemberService;
 import org.cotato.csquiz.domain.generation.entity.Generation;
@@ -40,47 +33,9 @@ public class AttendanceAdminService {
 
     private final GenerationReader generationReader;
     private final AttendanceRepository attendanceRepository;
-    private final AttendanceRecordRepository attendanceRecordRepository;
     private final AttendanceRecordService attendanceRecordService;
     private final SessionRepository sessionRepository;
     private final MemberService memberService;
-
-    @Transactional
-    public void addAttendance(Session session, Location location, LocalDateTime attendanceDeadline,
-                              LocalDateTime lateDeadline) {
-        AttendanceUtil.validateAttendanceTime(session.getSessionDateTime(), attendanceDeadline, lateDeadline);
-        if (session.hasOfflineSession()) {
-            checkLocation(location);
-        }
-        Attendance attendance = Attendance.builder()
-                .session(session)
-                .location(location)
-                .attendanceDeadLine(attendanceDeadline)
-                .lateDeadLine(lateDeadline)
-                .build();
-
-        attendanceRepository.save(attendance);
-    }
-
-    private void checkLocation(Location location) {
-        if (location == null) {
-            throw new AppException(ErrorCode.INVALID_LOCATION);
-        }
-    }
-
-    @Transactional
-    public void updateAttendanceRecords(Long attendanceId, Long memberId, AttendanceResult attendanceResult) {
-        Attendance attendance = attendanceRepository.findById(attendanceId)
-                .orElseThrow(() -> new EntityNotFoundException("해당 출석이 존재하지 않습니다"));
-
-        AttendanceRecord attendanceRecord = attendanceRecordRepository.findByMemberIdAndAttendanceId(memberId, attendanceId)
-                .orElseGet(() -> AttendanceRecord.absentRecord(attendance, memberId));
-
-        // Todo https://github.com/IT-Cotato/COTATO-BE/issues/204
-        attendanceRecord.updateAttendanceResult(attendanceResult);
-
-        attendanceRecordRepository.save(attendanceRecord);
-    }
 
     public List<GenerationMemberAttendanceRecordResponse> findAttendanceRecords(Long generationId) {
         List<Long> sessionIds = sessionRepository.findAllByGenerationId(generationId).stream().map(Session::getId).toList();
